requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
Pillow
openai
pdf2image
PyMuPDF